refactor(game): folder structure
