+++ src/game/game.tsx
@@ -1,9 +1,9 @@
 import { useApp } from '@inlet/react-pixi';
 import { Container } from 'pixi.js';
 import { useEffect, useState } from 'react';
-import { InitializeCelestials } from './generate';
-import { Star } from './star';
-import { useResize } from './use-resize.hook';
+import { Star } from './graphics/star';
+import { InitializeCelestials } from './utils/generate';
+import { useResize } from './utils/use-resize.hook';
 
 export const Game = () => {
   const app = useApp();
