+++ src/game/graphics/star.tsx
