+++ src/game/utils/generate.ts
