+++ src/game/utils/use-resize.hook.ts
@@ -1,5 +1,9 @@
 import { useEffect, useState } from 'react';
-import { footerHeight, sideNavWidth, topBarHeight } from '../components/layout';
+import {
+  footerHeight,
+  sideNavWidth,
+  topBarHeight,
+} from '../../components/layout';
 
 const getScreenSize = () => ({
   width: window.innerWidth - sideNavWidth,
